feat: optimize filter text calculation
